requirements.txt
numpy==2.1.3  
tqdm==4.67.0
pandas==2.2.3 
geopandas==1.0.1
geotiff==1.7.3
rasterio==1.4.3
opencv==4.10.0
matplotlib>=3.7
shapely>=2.0
pillow>=9.5 
